[CodeGen] Split off PseudoSourceValueManager into separate header (NFC)
Most users of PseudoSourceValue.h only need PseudoSourceValue,
not the PseudoSourceValueManager. However, this header pulls in
some very expensive dependenceis like ValueMap.h, which is only
used for the manager.

Split off the manager into a separate header and include it only
where used.
